Interface/requirements.txt
# Interface gráfica
flet

# MQTT sobre TLS (HiveMQ Cloud)
paho-mqtt

# Planilha .xlsx dos favoritos (a GUI cria/atualiza)
openpyxl
